requirements.txt
Flask>=2.0.0,<2.3.0
Flask-SQLAlchemy>=2.5.0,<3.0.0
Flask-Login>=0.5.0,<0.6.0
Flask-WTF>=1.0.0,<1.1.0
Flask-Migrate>=3.0.0,<3.1.0
Flask-Caching>=1.10.0,<2.0.0
Flask-Limiter>=2.0.0,<2.1.0
Flask-Mail>=0.9.1,<0.10.0
Werkzeug>=2.0.0,<2.1.0
SQLAlchemy>=1.4.0,<1.5.0
python-dotenv>=0.19.0,<0.20.0
qrcode>=7.3.0,<7.4.0
python-dateutil>=2.8.2,<2.9.0
pytz>=2021.3,<2022.0
gunicorn>=20.1.0,<20.2.0
requests>=2.27.0,<2.28.0
psycopg2-binary>=2.9.3,<2.10.0
Flask-Session>=0.4.0,<0.5.0
Pillow>=8.0.0,<9.0.0
numpy
pandas
openpyxl>=3.0.9,<3.1.0
dnspython>=2.2.0,<2.3.0
email_validator>=1.1.0,<1.2.0
